// PM_Case_Management_2/client/src/app/pages/Case/case-detail/make-appointment-case/Iappointmentwithcalander.ts
export interface IAppointmentWithCalander {


    EmployeeId: string
    AppointementDate: string
    Time: string
    CaseId: string
    CreatedBy: string
}

export interface IAppointmentGet {


    id: string
    name: string
    description: string
    badge: string
    date: string
    type: string
    everyYear: boolean
}